lib: bin: sb_fota: Add library for Softbank FOTA
This is a Softbank specific FOTA library, which is intented
to work in SB network as they have specified in separate document
and it is required in their carrier network.

This is a initial release of carrier specific library which
is not yet certified, so must be assumed to be work in progress.

Signed-off-by: Seppo Takalo <seppo.takalo@nordicsemi.no>