net: lib: lwm2m_client_utils: add Kconfig for security settings
Client utilities is handling server name and settings of sec_tag,
so these settings should be on its Kconfig, instead of relying
CONFIG_APP ones.

Reference to PSK key is dropped from the library, as application may
choose to write it, or not. It can be already on the modem.

Signed-off-by: Seppo Takalo <seppo.takalo@nordicsemi.no>